refactor: file 파일 구조 재정리(#47) - File 폴더 안에 File.jsx, File.css 옮겼다 - file-table의 row를 따로 FileItem component 뺐다 - 파일 선택, 파일 다운로드 훅들을 hooks 폴더에 넣었다
