+++ src/routes/File/File.css
@@ -22,6 +22,7 @@
 .file-table tr {
   border: none;
 }
+
 .check-icon {
   width: 24px;
   height: 24px;
@@ -40,10 +41,12 @@
   white-space: nowrap;
   text-overflow: ellipsis;
 }
-.file-table tbody td:nth-child(1){
+
+.file-table tbody td:nth-child(1) {
   text-align: center;
 }
-.file-table th:nth-child(3){
+
+.file-table th:nth-child(3) {
   text-align: center;
 }
 
@@ -58,65 +61,20 @@
   white-space: nowrap;
   text-overflow: ellipsis;
 }
-.file-table th:nth-child(1),
-.order-column {
+
+.file-table th:nth-child(1) {
   width: 7%;
   text-align: center;
   position: relative;
 }
 
-.order-column span {
-  position: absolute;
-   top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%); 
-}
 
-.info-column {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-.info-column div {
-  line-height: 46px;
-  height: 46px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-.category-col {
-  flex: 0 0 20%;
-  text-align: center;
-}
-
-.title-col {
-  flex: 0 0 50%;
-}
-
-.date-col {
-  flex: 0 0 30%;
-}
-
-.file-table th:nth-child(3),
-.download-column {
+.file-table th:nth-child(3) {
   width: 10%;
   position: relative;
 }
 
-.download-btn {
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%); 
-}
 
-.category {
-  padding: 6px 10px;
-  border-radius: 10px;
-}
 
 @media screen and (max-width: 900px) {
   .extend-file-menu {
@@ -124,13 +82,14 @@
     z-index: 899;
     width: 100%;
     top: 0;
-    left: 0; 
+    left: 0;
     background-color: #193553;
     border-radius: 0px 0px 16px 16px;
-    padding: 10px 25px 10px 25px; 
+    padding: 10px 25px 10px 25px;
     padding-top: 58px;
-    box-sizing: border-box; 
+    box-sizing: border-box;
   }
+
   .file-main-inner {
     padding-left: 0;
     margin: 10px auto;
@@ -149,53 +108,33 @@
     display: flex;
     align-items: center;
   }
+
   .file-table tbody th,
   .file-table tbody td {
     padding: 15px 0;
   }
 
-  .order-column {
-    width: 15%;
-    text-align: center;
-  }
-
-  .info-column {
-    width: 70%;
-    padding: 0 10px;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .download-column {
-    width: 15%;
-    text-align: center;
-  }
-
-  .category-col {
-    text-align: left;
-  }
-  .info-column div {
-    line-height: 32px;
-    height: 32px;
-  }
-  .extend-file-menu h2{
+  .extend-file-menu h2 {
     color: #fff;
     font-size: 14px;
     margin: 0;
     padding: 0;
     font-weight: 400;
   }
-  .extend-file-header{
+
+  .extend-file-header {
     display: flex;
     flex-direction: row;
     align-items: center;
   }
-  .extend-file-header img{
+
+  .extend-file-header img {
     width: 20px;
     height: 20px;
     cursor: pointer;
   }
-  .extend-file-menu-content{
+
+  .extend-file-menu-content {
     padding: 0 20px;
   }
 }
