+++ src/routes/File/components/FileItem.css
@@ -1,0 +1,84 @@
+.order-column {
+  width: 7%;
+  text-align: center;
+  position: relative;
+}
+
+.order-column span {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.info-column {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+.info-column div {
+  line-height: 46px;
+  height: 46px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+.category-col {
+  flex: 0 0 20%;
+  text-align: center;
+}
+
+.title-col {
+  flex: 0 0 50%;
+}
+
+.date-col {
+  flex: 0 0 30%;
+}
+
+.download-column {
+  width: 10%;
+  position: relative;
+}
+
+.download-btn {
+  width: 24px;
+  height: 24px;
+  cursor: pointer;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.category {
+  padding: 6px 10px;
+  border-radius: 10px;
+}
+
+@media screen and (max-width: 900px) {
+  .order-column {
+    width: 15%;
+    text-align: center;
+  }
+
+  .info-column {
+    width: 70%;
+    padding: 0 10px;
+    display: flex;
+    flex-direction: column;
+  }
+
+  .download-column {
+    width: 15%;
+    text-align: center;
+  }
+
+  .category-col {
+    text-align: left;
+  }
+  .info-column div {
+    line-height: 32px;
+    height: 32px;
+  }
+}
